requirements.txt
flask
google-genai
transformers
torch
sentence-transformers
numpy
opencv-python
scipy
ffmpeg-python
pillow
tqdm
openai-whisper
librosa
scenedetect
anthropic
openai
ollama
python-dotenv
